Removed unnecessary user commands and made user commands top-level
