+++ lib/codecitycli/user.rb
@@ -70,7 +70,11 @@ module CodeCityCLI
     end
 
     def authenticate
-      @token = AuthToken.new(Request.post(auth_path, { email: @email, password: @password }))
+      response = Request.post(auth_path, { email: @email, password: @password })
+      if response[:body].key?(:success) and response[:body][:success] == false
+        raise AuthenticationError, 'invalid login credentials'
+      end
+      @token = AuthToken.new(response)
     end
   end
 
@@ -95,65 +99,6 @@ module CodeCityCLI
   class Admin < Account
     def auth_path
       '/admins/sign_in'
-    end
-  end
-
-  module CLI
-    class User < Thor
-      desc 'login EMAIL PASSWORD', 'log into code city with EMAIL and PASSWORD'
-      option :as, type: :string, required: true
-      def login(email, password)
-        user = CodeCityCLI::User.new(email: email, password: password, user_type: options[:as].to_sym)
-
-        user.account.authenticate
-        token = user.account.token
-
-        Config.instance.token = token
-        Config.instance.user_id = user.account.id
-        Config.instance.user_type = user.user_type
-        Config.instance.save
-      end
-
-      desc 'new EMAIL PASSWORD', 'creates a new user with EMAIL and PASSWORD and signs it in'
-      option :user_type, type: :string, default: 'student'
-      option :first_name, type: :string, required: true
-      option :last_name, type: :string, required: true
-      def new(email, password)
-        user = CodeCityCLI::User.new(user_type: options[:user_type], first_name: options[:first_name], last_name: options[:last_name], email: email, password: password)
-
-        token = user.create
-
-        Config.instance.api_key = token
-        Config.instance.user_id = user.id
-        Config.instance.save
-      end
-
-      desc 'update', 'updates the current user'
-      option :email, type: :string
-      option :password, type: :string
-      option :user_type, type: :string
-      option :first_name, type: :string
-      option :last_name, type: :string
-      def update
-        user = CodeCityCLI::User.current_user
-
-        user.email = options[:email] if options[:email]
-        user.password = options[:password] if options[:password]
-        user.user_type = options[:user_type] if options[:user_type]
-        user.first_name = options[:first_name] if options[:first_name]
-        user.last_name = options[:last_name] if options[:last_name]
-
-        user.update
-      end
-
-      desc 'delete', 'deletes the current user'
-      def delete
-        user = CodeCityCLI::User.current_user
-
-        user.delete
-
-        Config.instance.api_key = nil
-      end
     end
   end
 end
